pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vedic-ai"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "PyYAML>=6.0",
    "typer>=0.12",
    "structlog>=24.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-snapshot",
]
engine = [
    "flatlib",
    "kerykeion",
]
retrieval = [
    "sentence-transformers",
    "faiss-cpu",
    "chromadb",
]
llm = [
    "httpx>=0.27",
    "json5",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
]

[project.scripts]
vedic-ai = "vedic_ai.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/vedic_ai"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
src = ["src"]
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
